Refactoring into contextrouter & change of default app architectures to OOP
This commit introduces several breaking changes in the API to improve architecture and stability
a. Refactored server as two packages - contextrouter and server for upcoming "noserver" option
b. Changed the use of context.Context to purely request scoped values as advised in the context package
c. Changed the architecture of the default example apps to an architecture more compatible with Java OOP
d. This architecture change also removes package globals